requirements.txt
# Framework principal
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4

# Autenticação
Flask-JWT-Extended==4.5.2
PyJWT==2.8.0
werkzeug==2.3.7

# Banco de dados
SQLAlchemy==2.0.19
psycopg2-binary==2.9.9

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3

# Machine Learning/Visão Computacional
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196  # YOLOv8
scikit-learn==1.3.0

# Processamento de imagens
Pillow==10.0.1

# Testes (opcional para produção)
pytest==7.4.2